requirements.txt
binance-connector==3.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
numpy==1.24.4
pandas==2.0.3
pandas-ta==0.3.14b0
pycryptodome==3.18.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing==3.7.4.3
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.6.2
